PoC for Ramp integration
