+++ app/javascript/components/Ramp.scss
@@ -1,0 +1,13 @@
+.iiif-player {
+  .ramp--structured-nav {
+    float: left;
+    width: 40%;
+    margin-top: 20px;
+  }
+  .ramp--transcript_nav {
+    float: right;
+    width: 60%;
+    margin-left: -20px;
+    margin-top: 20px;
+  }
+}
