+++ app/views/media_objects/_item_view.html.erb
@@ -29,20 +29,13 @@ Unless required by applicable law or agreed to in writing, software distributed
     <% if lending_enabled?(@media_object) && !can_stream %>
       <%= render 'embed_checkout' %>
     <% elsif @currentStream %>
-      <%= render partial: "modules/player/section", locals: {section: @currentStream, section_info: @currentStreamInfo, f_start: @f_start, f_end: @f_end} %>
-      <%= render file: '_track_scrubber.html.erb' if is_mejs_2? %>
-      <%= render file: '_add_to_playlist.html.erb' if current_user.present? && is_mejs_2? %>
-      <%# Partial view for MEJS4 Add To Playlist plugin %>
-      <%= render partial: 'mejs4_add_to_playlist' if current_user.present? && is_mejs_4? %>
       <%= render 'workflow_progress' %>
       <%= render partial: 'timeline' if current_ability.can? :create, Timeline %>
       <%= render 'share' if will_partial_list_render? :share %>
-
-      <!-- Sections -->
-      <%= render partial: 'sections',
-          locals: { mediaobject: @media_object,
-                  sections: @masterFiles,
-                  activeStream: @currentStream } %>
+      <% if can_stream && @currentStream %>
+        <% supplemental_files = @masterFiles.map { |mf| {"id" => mf.id, "transcripts" => mf.supplemental_files.select { |sf| sf.tags.include?('transcript')} } }.flatten(1) %>
+        <%= react_component("Ramp", {base_url: request.protocol+request.host_with_port, transcripts: supplemental_files, mo_id: @media_object.id }) %>
+      <% end %>
     <% end %>
   </div>
   <div class="col-sm-4" id="right-column">
