+++ package.json
@@ -4,6 +4,7 @@
     "@babel/preset-react": "^7.0.0",
     "@rails/webpacker": "^5.2.2",
     "@samvera/iiif-react-media-player": "1.2.1",
+    "@samvera/ramp": "./samvera-ramp-2.0.1-avalon-test.tgz",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "cropperjs": "^1.5.5",
     "hls.js": "https://github.com/avalonmediasystem/hls.js#stricter_ts_probing",
@@ -15,7 +16,7 @@
     "react-structural-metadata-editor": "https://github.com/avalonmediasystem/react-structural-metadata-editor",
     "react_ujs": "^2.4.4",
     "url-search-params-polyfill": "^7.0.1",
-    "video.js": "^7.15.4",
+    "video.js": "7.21.4",
     "videojs-hotkeys": "^0.2.27"
   },
   "devDependencies": {
